Rename files in the compiler server
This is the final part of the type hierarchy unwind. It renames the
files in the projects to match the new type names.
