+++ src/Compilers/Server/VBCSCompilerTests/NamedPipeClientConnectionTests.cs
